+++ CHANGELOG.md
@@ -8,7 +8,7 @@ All notable changes to this project will be documented in this file.
 
 - [matterbridge]: added unregisterAllDevices() to the platforms 
 - [matterbridge]: added unregisterDevice(device: MatterbridgeDevice) to the platforms
-- [frontend]: Enable and disable plugin are now available. The process restart.
+- [frontend]: Enable and disable plugin are now available. Restart Matteerbridge after.
 - [frontend]: Frontend got updated to 0.8.2.
 
 ## [1.1.9] - 2024-03-15